new overall Maven configuration. Main change is that the IDE now uses only the maven eclipse plug-in. A lot of unused configurations have also been removed (mainly because they are inherited from super poms and dont need to be repeated).
The IDE can now be build from the ide/build artifact by deploy which produces target/RCP with all platforms *Required an Eclipse installation with delta package and settings configuration in .m2

The new Maven configuration supports Maven 3.

General setup instructions are:
1. check out trunk
2. core/ run mvn install
3. ide/ run mvn install -P nocompile
4. trunk run mvn eclipse:eclipse
5. ide/ run mvn eclipse:eclipse

Now the projects can be imported in Eclipse through the import existing Maven projects. The reason for the repeated eclipse:eclipse is that it second one uses the .classpath to sync the runtime config in the MANIFEST.MF.

git-svn-id: https://overture.svn.sourceforge.net/svnroot/overture/trunk@3509 2b2d4d09-c93c-414a-ad04-6e973f5c1cff